requirements.txt
pandas==1.5.3
numpy==1.24.2
scikit-learn==1.2.2
matplotlib==3.7.2
seaborn==0.12.2
plotly==5.15.0
pyyaml==6.0
pyod==1.1.0
yara-python==4.2.3
tqdm==4.64.1
